Add endpoint for deleting event

+++ Models/Event.cs
@@ -6,7 +6,7 @@ public class Event
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string Location { get; set; } = string.Empty;
-    public decimal PricePerPerson { get; set; }
+    public float PricePerPerson { get; set; }
     public DateTime Date { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.Now;
     public DateTime UpdatedAt { get; set; } = DateTime.Now;
